Move functionality from Game to Computer
