+++ battle_ship_runner.rb
@@ -2,6 +2,8 @@ require './lib/board'
 require './lib/ship'
 require './lib/game'
 require './lib/cell'
+require './lib/player'
+require './lib/computer'
 
 game = Game.new
 game.welcome
