+++ lib/player.rb
@@ -58,6 +58,16 @@ class Player
     user_submarine_place
   end
 
+
+  def user_submarine_place
+    submarine_greeting
+    submarine_place
+  end
+
+  def prompt_shot
+    puts "Enter your coordinate for your shot!!"
+  end
+
   def submarine_greeting
     puts
     puts "Now place your submarine. Don't place it on your cruiser."
@@ -89,6 +99,26 @@ class Player
       end
       # create_ship
     end
+
+    # def player_shot
+    #   prompt_shot
+    #   shot = gets.chomp.upcase!
+    #   if computer_board.valid_coordinate?(shot) && computer_board.cells[shot].fired_upon?
+    #     puts
+    #     puts "!!!Already fired here!!!"
+    #     puts "------------------------"
+    #     player_shot
+    #   elsif computer_board.valid_coordinate?(shot) && computer_board.cells[shot].fired_upon? == false
+    #     computer_board.fire(shot)
+    #   else
+    #     puts
+    #     puts "That's an invalid coordinate. What are you doing?"
+    #     puts "^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^"
+    #     player_shot
+    #   end
+    #   puts
+    #   show_computer_board
+    # end
